Billing-Team/requirements.txt
mysql-connector-python==9.1.0
orm-mysql==1.0.6
PyYAML==6.0.2
